Added in-game message for restart status Added in missing value for $STOP_MSG variable
I felt like the users in-game should know whether or not the server would be coming back up soon, so I added a check for the status of the service, and then upon success had a message sent to the bukkit console before the mc_stop function is called.

Also, I forgot to add in a message for the STOP_MSG variable, so I did so. Decided to add in a 5 sec delay before doing a save-all to give in-game users more of a heads up.