Bumped version number to 0.2.1
Should have been bumped to 2.0 earlier when the ramdisk functions were added, but it's not a big deal. Added 0.0.1 for the last commit (minor changes).

Version 1.0.0 will identify the script as feature-complete (that is, 100% working and reliable ramdisk and backup functions)

For future reference, in a.b.c, "a" is a major version, signifying significant changes. "b" is a minor version signifying added features. "c" is for minor tweaks and bug-fixes. An added "-x" where x = some number will be added for hotfixes if needed (i.e. 1.3.2-1).